RagePixel and Camera
I added the RagePixel plugin to the project. I also added camera
functionality that adapts to the size of the grid. The camera centers in
the middle of the grid and zooms in or out depending on how big the map
is so that the entire map is visible at all times.